Moved language selection to input area
Removed separate system level message that acted as language selection bubble. Language selection happens now in the input area.
